a re-organization of header code for eigen, mpfr extensions
particularly related to boost multiprecision extensions and eigen extensions.  this re-organization was spurred by the conversation with John Maddock and chtz regarding Boost multiprecision bug 11149, where min and max are missing for expression templated number types.  the bug shows up when attempting to take min or max of such types, and first showed up for us when Jeb was working on exposing eigen matrix operations for mpfr types through boost multiprecision.

this commit does not solve the problem, specifically, that svd doesn't compile when using mpfr_float with et_on.  rather, it allows to use et_on with min/max without the (max)() syntax.  furthermore, this commit helps make the contents of some header files more accurately placed.

when including any eigen headers, one should include eigen_extensions.hpp first.
