+++ core/include/bertini2/tracking/euler.hpp
@@ -34,6 +34,7 @@
 #include "tracking/tracking_config.hpp"
 
 #include "system.hpp"
+#include "mpfr_extensions.hpp"
 #include <Eigen/LU>
 
 #include <boost/type_index.hpp>
